Cleans out Comments for streamlining a new plugin
Renames the main php file to 'plugin.php'
Imports the name of the block from the block.json and uses it in the registration. This allows the block.json to be the single source of information for the plugin, and we don't have to update the name in multiple locations.